Carry the row identifier at its full width into an R-tree node
rtree_insert takes an int64 identifier and a node stores int64, but node_insert
between them took an int, so an identifier of 2^31 or more was truncated on the
way in. The entry is stored and found correctly and only the number reported
for it changes, so a caller that fetches by the returned identifier reads a
different row, or none.

Every count and every window answer stays right, which is why nothing caught
it. The R-tree tests all insert the loop variable as the identifier, and
rtree_span_test uses it as an array subscript, so none of them could carry a
value that large.

test_id_width inserts three identifiers above 2^31, one of them 2^40, and
requires every identifier the search reports to be one that was inserted.
Against the narrower carrier it reports -294967296 for 4000000000.

+++ meos/src/temporal/temporal_rtree.c
@@ -727,7 +727,7 @@ node_split(RTree *rtree, RTreeNode *node, void *box, RTreeNode **right_out)
  */
 static void
 node_insert(RTree *rtree, void *node_bounding_box, RTreeNode *node,
-  void *new_box, int id, bool *split)
+  void *new_box, int64 id, bool *split)
 {
   if (node->node_type == RTREE_LEAF)
   {
